splitting files out.
